requirements.txt
asgiref==3.7.2
Django==4.2.11
django-debug-toolbar==3.2.4
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
pytz==2024.1
requests==2.7.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
jwt==1.3.1